CRUD operasyonları yazıldı

+++ Entities/DTOs/CarDetailDto.cs
@@ -12,6 +12,6 @@ namespace Entities.DTOs
         public int CarId { get; set; }
         public string CarName { get; set; }
         public string BrandName { get; set; }
-        public double DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
     }
 }
